requirements.txt
langchain-openai
streamlit
python-dotenv
numpy
pandas
langchain-community